Generalize non-rigorous dual bound detection across GDPopt and MindtPy (#3947)
Extract the convexity/curvature classification out of GDP_LOA_Solver into a
shared pyomo/contrib/gdpopt/nonrigorous_bounds.py module, and drive it from a
single _crossed_bounds_are_certified class flag on both the GDPopt and MindtPy
algorithm base classes.

An algorithm is certified when its relaxation is valid for any model it accepts.
The McCormick-based global algorithms (GDPopt GLOA, MindtPy GOA) are certified.
The algorithms that linearize at trial points (GDPopt LOA, MindtPy OA and ECP)
are only valid on convex models, so they are not certified and the model is
checked for convexity before crossed bounds are trusted.

This corrects the MindtPy classification, which previously marked GOA (the
McCormick-based algorithm) as uncertified while leaving OA and ECP certified,
and adds MindtPy the model-structure gate that GDPopt already had.
