+++ pyomo/contrib/mindtpy/global_outer_approximation.py
@@ -35,7 +35,8 @@ class MindtPy_GOA_Solver(_MindtPyAlgorithm):
     """
 
     CONFIG = _get_MindtPy_GOA_config()
-    _crossed_bounds_are_certified = False
+    # GOA relaxes the nonlinear constraints with McCormick envelopes, which is a
+    # valid relaxation of nonconvex problems, so its dual bound is rigorous.
 
     def check_config(self):
         config = self.config
